refactor: only warn for config issues once
